TT-1: Initial set up Spring Context and Tests
1) Set up spring-app.xml
2) Set up Tests
3) Add simple test
