--- IncorrectFormatExpression.java
@@ -1,7 +1,0 @@
-package com.byndyusoft;
-
-public class IncorrectFormatExpression extends RuntimeException {
-    public IncorrectFormatExpression(String message) {
-        super(message);
-    }
-}
