+++ src/main/java/com/byndyusoft/IncorrectFormatExpression.java
@@ -1,0 +1,7 @@
+package com.byndyusoft;
+
+public class IncorrectFormatExpression extends RuntimeException {
+    public IncorrectFormatExpression(String message) {
+        super(message);
+    }
+}
