+++ src/test/java/com/byndyusoft/AbstractTest.java
@@ -1,0 +1,8 @@
+package com.byndyusoft;
+
+import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
+
+@SpringJUnitConfig(locations = {"classpath:spring/spring-app.xml"})
+public class AbstractTest {
+
+}
